Refactored code to follow a naming convention for firebase stuff
