+++ app/src/main/java/com/example/habitize/AddHabitActivity.java
@@ -7,14 +7,20 @@ import android.widget.Button;
 import android.widget.EditText;
 import android.widget.Switch;
 
+import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
 
 import com.google.firebase.firestore.CollectionReference;
 import com.google.firebase.firestore.DocumentReference;
+import com.google.firebase.firestore.DocumentSnapshot;
+import com.google.firebase.firestore.EventListener;
 import com.google.firebase.firestore.FirebaseFirestore;
+import com.google.firebase.firestore.FirebaseFirestoreException;
 
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 public class AddHabitActivity extends AppCompatActivity {
     private EditText title;
@@ -33,7 +39,7 @@ public class AddHabitActivity extends AppCompatActivity {
     private Button createHabit;
     private User currentUser;
     private FirebaseFirestore db;
-    private CollectionReference users;
+    private CollectionReference userCol;
     private DocumentReference docRef;
     private String passedEmail;
     private List<Habit> passedHabits;
@@ -42,7 +48,6 @@ public class AddHabitActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_add_habit_info);
         passedEmail = (String)getIntent().getExtras().getSerializable("User"); // retrieving passed user
-        passedHabits = (List<Habit>)getIntent().getExtras().getSerializable("list");
 
         imageBtn = findViewById(R.id.addImage);
         locationBtn = findViewById(R.id.addLocation);
@@ -50,11 +55,27 @@ public class AddHabitActivity extends AppCompatActivity {
         title = findViewById((R.id.habitTitle));
         description = findViewById((R.id.habitDescription));
         db = FirebaseFirestore.getInstance();
-        users = db.collection("userHabits");
-        docRef = users.document(passedEmail);
+        userCol = db.collection("userHabits");
+        docRef = userCol.document(passedEmail);
         //TO DO: make a header w/ title - NEW HABIT - perhaps textview trick
 
 
+        docRef.addSnapshotListener(new EventListener<DocumentSnapshot>() {
+            @Override
+            public void onEvent(@Nullable DocumentSnapshot value, @Nullable FirebaseFirestoreException error) {
+                passedHabits = new ArrayList<>(); // reset the list
+                ArrayList<Habit> mappedList =  (ArrayList<Habit>) value.get("habits");
+                for(int i = 0; i < mappedList.size() ; i++){ // get each item one by one
+                    Map<String,String> habitFields = (Map<String, String>) mappedList.get(i); // map to all the fields
+                    // retrieves all the habit information and adds it to the habitList
+                    String name = habitFields.get("name");
+                    String description = habitFields.get("description");
+                    Habit newHabit = new Habit(name,description); // create a new habit out of this information
+                    passedHabits.add(newHabit); // add it to the habitList
+
+                }
+            }
+        });
 
         imageBtn.setOnClickListener(new View.OnClickListener() {
             @Override
