+++ app/src/main/java/com/example/habitize/MainActivity.java
@@ -44,7 +44,7 @@ public class MainActivity extends AppCompatActivity {
     private ProgressBar progressBar3;
     private TextView textView2;
     private Toolbar toolBar;
-    private String currentUser;
+    private String passedEmail;
     private CollectionReference progress;
     private CollectionReference userHabits;
     private DocumentReference user;
@@ -64,28 +64,13 @@ public class MainActivity extends AppCompatActivity {
         logOut = findViewById(R.id.logout);
         db = FirebaseFirestore.getInstance();
         userHabits = db.collection("userHabits");
-        currentUser = (String)getIntent().getExtras().getSerializable("User"); // retrieving the user
-        user = userHabits.document(currentUser); // gets the habits at the current user
+        passedEmail = (String)getIntent().getExtras().getSerializable("User"); // retrieving the user
+        user = userHabits.document(passedEmail); // gets the habits at the current user
 
 
 
 
-        user.addSnapshotListener(new EventListener<DocumentSnapshot>() {
-            @Override
-            public void onEvent(@Nullable DocumentSnapshot value, @Nullable FirebaseFirestoreException error) {
-                habitList = new ArrayList<>(); // reset the list
-                ArrayList<Habit> mappedList =  (ArrayList<Habit>) value.get("habits");
-                for(int i = 0; i < mappedList.size() ; i++){ // get each item one by one
-                    Map<String,String> habitFields = (Map<String, String>) mappedList.get(i); // map to all the fields
-                    // retrieves all the habit information and adds it to the habitList
-                    String name = habitFields.get("name");
-                    String description = habitFields.get("description");
-                    Habit newHabit = new Habit(name,description); // create a new habit out of this information
-                    habitList.add(newHabit); // add it to the habitList
 
-                }
-            }
-        });
 
         addHabit = findViewById(R.id.addHabit); // our 4 buttons
         allHabits = findViewById(R.id.allHabits);
@@ -111,7 +96,7 @@ public class MainActivity extends AppCompatActivity {
                 // passsing down a list to modify, and a user to change the database of after the change is made
                 Bundle userBundle = new Bundle();
                 userBundle.putSerializable("list",habitList);
-                userBundle.putSerializable("User",currentUser);
+                userBundle.putSerializable("User",passedEmail);
                 intent.putExtras(userBundle);
                 startActivity(intent);
 
@@ -130,7 +115,7 @@ public class MainActivity extends AppCompatActivity {
                 // passing list down to populate listView
                 Bundle userBundle = new Bundle();
                 userBundle.putSerializable("list",habitList);
-                userBundle.putSerializable("user",currentUser);
+                userBundle.putSerializable("user",passedEmail);
                 intent.putExtras(userBundle);
                 startActivity(intent);
